Move ts constants file out of cbind folder.
