+++ yarn-project/circuits.js/src/index.ts
@@ -4,4 +4,4 @@ export * from './rollup/index.js';
 export * from './utils/jsUtils.js';
 export * from './contract/index.js';
 export * from './types/index.js';
-export * from './cbind/constants.gen.js';
+export * from './constants.gen.js';
