+++ yarn-project/circuits.js/src/scripts/constants.in.ts
@@ -4,7 +4,7 @@ import * as fs from 'fs';
 import { dirname, join } from 'path';
 
 const NOIR_CONSTANTS_FILE = '../../../aztec-nr/aztec/src/constants_gen.nr';
-const TS_CONSTANTS_FILE = './constants.gen.ts';
+const TS_CONSTANTS_FILE = '../constants.gen.ts';
 const SOLIDITY_CONSTANTS_FILE = '../../../../l1-contracts/src/core/libraries/ConstantsGen.sol';
 
 /**
