+++ yarn-project/circuits.js/src/structs/rollup/base_or_merge_rollup_public_inputs.ts
@@ -1,7 +1,7 @@
 import { Fr } from '@aztec/foundation/fields';
 import { BufferReader } from '@aztec/foundation/serialize';
 
-import { NUM_FIELDS_PER_SHA256 } from '../../cbind/constants.gen.js';
+import { NUM_FIELDS_PER_SHA256 } from '../../constants.gen.js';
 import { serializeToBuffer } from '../../utils/serialize.js';
 import { AggregationObject } from '../aggregation_object.js';
 import { RollupTypes } from '../shared.js';
